+++ src/components/app-bar/menu/title.container.jsx
@@ -2,51 +2,14 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 // import { connect } from 'react-redux';
-import { withStyles } from '@material-ui/core/styles';
+import theme from '@app/theme'
 import Typography from '@material-ui/core/Typography';
-
+import { makeStyles } from '@material-ui/core';
 // TODO
 // @connect(state => ({
 //   settings: state.get('application').toJS().settings,
 // }))
-class ApplicationTitleContainer extends React.Component {
-  state = {
-    // auth: true,
-    // anchorEl: null,
-  };
-
-  render() {
-    // TODO
-    // const { settings } = this.props;
-    const settings = {
-      navbarTitle: 'Fix the title',
-      titleText: 'fix its text',
-    };
-    const { classes } = this.props;
-    return (
-      <Link to={settings.navbarLink} className={classes.title}>
-        <Typography
-          variant='h6'
-          className={classes.titleText}
-          component='h2'
-          noWrap
-        >
-          {settings.navbarTitle}
-        </Typography>
-      </Link>
-    );
-  }
-}
-
-ApplicationTitleContainer.propTypes = {
-  // items: PropTypes.array,
-};
-
-ApplicationTitleContainer.defaultProps = {
-  // items: [],
-};
-
-const style = (theme) => ({
+const useStyle = () => makeStyles({
   title: {
     fontSize: '1.5rem',
     display: 'block',
@@ -60,4 +23,39 @@ const style = (theme) => ({
   },
 });
 
-export default withStyles(style)(ApplicationTitleContainer);
+const ApplicationTitleContainer = props => {
+  // state = {
+  //   // auth: true,
+  //   // anchorEl: null,
+  // };
+
+  // TODO
+  // const { settings } = this.props;
+  const settings = {
+    navbarTitle: 'Fix the title',
+    titleText: 'fix its text',
+  };
+  const classes = useStyle();
+  return (
+    <Link to={settings.navbarLink} className={classes.title}>
+      <Typography
+        variant='h6'
+        className={classes.titleText}
+        component='h2'
+        noWrap
+      >
+        {settings.navbarTitle}
+      </Typography>
+    </Link>
+  );
+};
+
+ApplicationTitleContainer.propTypes = {
+  // items: PropTypes.array,
+};
+
+ApplicationTitleContainer.defaultProps = {
+  // items: [],
+};
+
+export default ApplicationTitleContainer;
