+++ src/pages/UniVerse/InputForm/predefined-patterns.js
