+++ src/pages/UniVerse/Result/beyt-card.jsx
