+++ src/pages/UniVerse/Result/error-card.jsx
