+++ src/pages/UniVerse/api/index.js
@@ -1,8 +1,8 @@
 import { getRequest } from '@utils/request';
 import { BASE_PATH } from '@app-settings';
 
-export const generatePoemRequest = ({ style, mask }) => {
-  getRequest({
+export const generatePoemRequest = async ({ style, mask }) => {
+  await getRequest({
     url: `${BASE_PATH}/deepsher/${style}/${mask}`,
   });
 };
